Cache per file, instead of all in one file

+++ LayoutEngine/Extensions.cs
@@ -90,15 +90,15 @@ namespace JBSnorro
 		/// <summary>
 		/// Quick and dirty get hash code implementation taking into account contents of the byte array.
 		/// </summary>
-		public static nint ComputeHash(this byte[]? data)
+		public static nuint ComputeHash(this byte[]? data)
 		{
 			if (data == null)
 			{
 				return 0;
 			}
 
-			nint i = data.Length;
-			nint hc = i + 1;
+			int i = data.Length;
+			nuint hc = (nuint)i + 1;
 
 			while (--i >= 0)
 			{
@@ -111,7 +111,7 @@ namespace JBSnorro
 		/// <summary>
 		/// Quick and dirty get hash code implementation taking into account contents of the string.
 		/// </summary>
-		public static nint ComputeHash(this string data)
+		public static nuint ComputeHash(this string data)
 		{
 			if (data == null)
 			{
@@ -119,7 +119,7 @@ namespace JBSnorro
 			}
 
 			int i = data.Length;
-			nint hc = i + 1;
+			nuint hc = (nuint)i + 1;
 
 			while (--i >= 0)
 			{
@@ -139,7 +139,7 @@ namespace JBSnorro
 		/// <summary>
 		/// Hashes the contents of the specified file.
 		/// </summary>
-		public static nint ComputeFileHash(this string path)
+		public static nuint ComputeFileHash(this string path)
 		{
 			var bytes = File.ReadAllBytes(path);
 			return bytes.ComputeHash();
@@ -147,7 +147,7 @@ namespace JBSnorro
 		/// <summary>
 		/// Hashes the contents of the specified file.
 		/// </summary>
-		public static async Task<nint> ComputeFileHashCodeAsync(this string path)
+		public static async Task<nuint> ComputeFileHashCodeAsync(this string path)
 		{
 			var bytes = await File.ReadAllBytesAsync(path);
 			return bytes.ComputeHash();
@@ -155,9 +155,9 @@ namespace JBSnorro
 		/// <summary>
 		/// Sums all specified numbers, without overflow exceptions.
 		/// </summary>
-		public static nint Sum(this IEnumerable<nint> summands)
+		public static nuint Sum(this IEnumerable<nuint> summands)
 		{
-			nint result = 0;
+			nuint result = 0;
 			foreach (var summand in summands)
 				result += summand;
 			return result;
